feat: 按照接口计划完成用户部分，商品部分，购物车部分的后端

+++ Backend/src/main/java/cn/airbozh/controller/UserController.java
@@ -3,9 +3,9 @@ package cn.airbozh.controller;
 import cn.airbozh.pojo.User;
 import cn.airbozh.service.UserService;
 import cn.airbozh.utility.Auth.Auth;
-import cn.airbozh.utility.JwtUtil;
+import cn.airbozh.utility.Auth.AuthException;
 import cn.airbozh.utility.ResMsg;
-import jdk.nashorn.internal.parser.Token;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
@@ -16,10 +16,10 @@ public class UserController {
     @Resource
     UserService userService;
     ResMsg<?> resMsg;
+
     @RequestMapping(value = "/account/{token}", method = RequestMethod.GET)
-    public ResMsg<?> isLogin(@PathVariable("token") String token){
-        String userId = JwtUtil.verifyToken(token).getBody().getSubject();
-        User user = userService.findUserById(Integer.parseInt(userId));
+    public ResMsg<?> isLogin(@PathVariable("token") String token) {
+        User user = userService.verifyToken(token);
         if (user == null) {
             resMsg = new ResMsg<>(0, "Already Login,But Find User Fail");
         } else {
@@ -31,8 +31,8 @@ public class UserController {
     @RequestMapping(value = "/account", method = RequestMethod.POST)
     public ResMsg<?> login(@RequestBody User user) {
         if (userService.userLogin(user)) {
-            int userId = userService.findUserByName(user.getUserName()).getUserId();
-            String token = JwtUtil.createToken(String.valueOf(userId));
+            User dbUser = userService.findUserByName(user.getUserName());
+            String token = userService.createToken(dbUser);
             resMsg = new ResMsg<>(1, "Login Success", token);
         } else {
             resMsg = new ResMsg<>(0, "Login Fail");
@@ -42,13 +42,15 @@ public class UserController {
 
     @RequestMapping(value = "/user", method = RequestMethod.GET)
     @Auth
-    public ResMsg<?> getUsers() {
+    public ResMsg<?> getUsers(User user) {
+        if (userService.findUserById(user.getUserId()).getRole() == 0)
+            throw new AuthException(HttpStatus.UNAUTHORIZED, "没有权限");
         List<User> userList = userService.getUsers();
-        if (userList == null) {
+        if (userList == null)
             resMsg = new ResMsg<>(0, "GetUsers Fail");
-        } else {
+        else
             resMsg = new ResMsg<>(1, "GetUsers Success", userList);
-        }
+
         return resMsg;
     }
 
@@ -65,37 +67,46 @@ public class UserController {
     }
 
     @RequestMapping(value = "/user", method = RequestMethod.POST)
-
     public ResMsg<?> registerUser(@RequestBody User user) {
-        if (userService.userRegister(user)) {
+        if (userService.userRegister(user))
             resMsg = new ResMsg<>(1, "Register Success");
-        } else {
+        else
             resMsg = new ResMsg<>(0, "Register Fail");
-        }
         return resMsg;
     }
 
     @RequestMapping(value = "/user/{userId}", method = RequestMethod.PATCH)
     @Auth
     public ResMsg<?> updateUser(@PathVariable("userId") int userId, @RequestBody User user) {
-        User dbUser = userService.userUpdate(userId, user);
-        if (dbUser != null)
-            resMsg = new ResMsg<>(1, "Update Success", dbUser);
-        else {
-            resMsg = new ResMsg<>(0, "Update Fail");
-        }
-        return resMsg;
+        if (user.getUserId() == userId || userService.findUserById(user.getUserId()).getRole() == 1) {
+            User dbUser = userService.userUpdate(userId, user);
+            if (dbUser != null)
+                resMsg = new ResMsg<>(1, "Update Success", dbUser);
+            else {
+                resMsg = new ResMsg<>(0, "Update Fail");
+            }
+            return resMsg;
+        } else
+            throw new AuthException(HttpStatus.UNAUTHORIZED, "没有权限");
     }
 
     @RequestMapping(value = "/user/{userId}", method = RequestMethod.DELETE)
     @Auth
-    public ResMsg<?> deleteUser(@PathVariable("userId") int userId) {
+    public ResMsg<?> deleteUser(@PathVariable("userId") int userId, User user) {
+        if (user.getUserId() != userId)
+            throw new AuthException(HttpStatus.UNAUTHORIZED, "没有权限");
         if (userService.deleteUser(userId)) {
             resMsg = new ResMsg<>(1, "DeleteUser Success");
         } else {
             resMsg = new ResMsg<>(0, "DeleteUser Fail");
         }
         return resMsg;
+    }
+
+    @RequestMapping(value = "/money", method = RequestMethod.GET)
+    @Auth
+    public ResMsg<?> getMoney(User user) {
+        return new ResMsg<>(1, "GetMoney Success", userService.getMoney(user.getUserId()));
     }
 
 }
